Bind reactor to thread
This commit introduces the default bind behavior of the reactor to have
its loop in a dedicated thread.

Previously, this was left for the user to decide, but then we might want
to avoid complicated schemes to interact with the created service such
as admin channels. Instead, the bind function will simply return the
allocated socket for the listener, and any interaction with the reactor
backend should be done from that entrypoint.