Decouple reactor bind from listen
This commit decouples the `listen` action from the `bind` address.

It is desirable to allow the user to fetch the used socket from the
listener and to bind the reactor socket as he wishes - either in a
dedicated worker thread or in the current main thread.

For type safety, a new type `ReactorListener` is introduced, and this
can be created only from `Reactor::bind`. The reactor will deref
non-mutable to tokio listener, and this will give freedom to the user to
query anything of relevance directly from the listener, without mutating
it.

When the user consumes the listener, an events loop will start in the
current thread. The user is free to move the listener to a dedicated
thread, or just listen in the current one.