Add custom add/remove custom breakpoint
The protocol will not allow stateless add/removal of breakpints because
`SetBreakpoint` requires the user to send the full set of breakpoints
instead of allowing them to be manipulated individually.

This is a questionable design since breakpoints can trivially be
referenced via their ids, that is available in the API.

This commit introduces a custom request that is not specified in the
protocol that will allow users to add and remove breakpoints without
keeping track of the state of the debugger.